+++ pkg/testutil/file.go
@@ -1,4 +1,4 @@
-package ops
+package testutil
 
 import (
 	"encoding/json"
